test(webhooks): move object-level authz tests in-src so --lib coverage counts them
The DB-backed object-level authorization tests lived in an integration
target (backend/tests/webhook_object_level_authz_tests.rs). CI computes
coverage with `cargo llvm-cov nextest --workspace --lib`, which only
instruments in-crate (`--lib`) tests, so those tests never counted the
new authz lines in webhooks.rs.

Relocate all 9 tests into an in-src `#[cfg(test)] mod
object_level_authz_tests` inside backend/src/api/handlers/webhooks.rs.
They call the handlers and the authz seam directly, acquire the pool via
test_db_helpers::try_pool(), and runtime-skip when DATABASE_URL is unset
(no `#[ignore]`), so CI -- which sets DATABASE_URL -- runs and counts
them. Delete the now-redundant integration file.

No production logic changed.
